增加主界面，颜色改用shell的方法，显示系统信息

+++ discover_log/view/login_view.py
@@ -6,15 +6,18 @@ import sys
 
 sys.path.append('..')
 from view.base_view import BaseView
-from discovery_log_file import get_all_log
-from termcolor import colored
+from discovery_log_file import get_all_log, calcMemUsage
+# from termcolor import colored
 import sqlite3
 
 # log文件显示行数
 line_show = 15
 log_list = ['nginx', 'httpd', 'mysqld', 'sys']
 
-line_help = 'q--返回上一页  退出o--输出  空格--进入  m--标记'
+line_help = 'q--返回上一页,退出  o--输出日志  空格--进入  m--标记  k--向下  j--向上'
+
+version = 'Linux&&Unix日志取证大师V1.0'
+
 
 class Login(BaseView):
     '''
@@ -23,7 +26,7 @@ class Login(BaseView):
 
     def __init__(self):
         super(Login, self).__init__()
-        self.title = '主界面'
+        self.title = '自定义选择界面'
         self.content = '请选择...'
         self.log_name = ''
         self.content_main = ['\r']
@@ -39,25 +42,45 @@ class Login(BaseView):
         self.mark = {'nginx': [], 'httpd': [], 'mysqld': [], 'sys': []}
         # 主页面的标记
         self.mark_main = []
+        # 是否正在进行数据库操作
+        self.working = False
+        # 
+        self._initSysInfo()
+        # self.memuse, self.memtotal = calcMemUsage()
+
+    def _initSysInfo(self):
+        memuse, memtotal = calcMemUsage()
+        import platform 
+        l = ['系统:' + str(platform.platform()) + '\r']
+        l.append('内存使用:' + str(memuse) + ' kb  内存总共:' + str(memtotal) + ' kb\r')
+        l.append('计算机类型:'+ str(platform.machine()) + '\r')
+        l.append('处理器信息:' + str(platform.processor()) + '\r')
+        self.sysinfo = l
 
     def make_displaylines(self):
         self.screen_height, self.screen_width = self.linesnum()
         display_lines = ['\r']
+        display_lines.append('\033[34m' + version + ' \033[0m' + '\r')
         display_lines.append(self.title + ' ' * 10 + self.log_name + '\r')
         display_lines.append('')
+        # display_lines.append('内存使用:' + str(self.memuse) + ' kb  内存总共:' + str(self.memtotal) + ' kb\r')
+        for i in self.sysinfo:
+            display_lines.append(i)
+        display_lines.append('')
+
         display_lines.append(self.content + str(self.select) + '\r')
         if not self.logshow:
             # 重置指针范围
             self.select_range = len(log_list) - 1
             for i in range(len(log_list)):
                 if i == self.select:
-                    line = '->' + ' ' * 10 + log_list[i]
+                    line = '->' + ' ' * 10 + '\033[42;37m' + log_list[i] + '\033[0m'
                     # display_lines.append('->' +' '*10 +log_list[i]+ '\r')
                 else:
                     line = ' ' * 12 + log_list[i]
                     # display_lines.append(' '*12 +log_list[i]+ '\r')
                 if i in self.mark_main:  # 标记
-                    line += colored("   ♡ ", 'red')
+                    line += '   \033[36m√ \033[0m'
                 line += '\r'
                 display_lines.append(line)
 
@@ -81,7 +104,7 @@ class Login(BaseView):
                 else:
                     line = ' ' * 12 + self.logfile[i]
                 if i in self.mark[self.log_name]:  # 标记
-                    line += colored("   ♡ ", 'red')
+                    line += '    \033[36m√ \033[0m'
                 line += '\r'
                 display_lines.append(line)
         #display_lines.append(' ' * 12 + str(self.select) + '\r')
@@ -111,35 +134,37 @@ class Login(BaseView):
         self.logfile = logfile
 
     def marking(self):
-        db = sqlite3.connect('loginfo.db')
-        cur = db.cursor()
-        if self.logshow:
-            filename = self.logfile[self.select]
-            print filename
-            if self.select in self.mark[self.log_name]:
-                self.mark[self.log_name].remove(self.select)
-                cur.execute('delete from log_select where name=%s' % filename)
-            else:
-                self.mark[self.log_name].append(self.select)
+        if not self.working:
+            self.working = True
+            db = sqlite3.connect('loginfo.db')
+            cur = db.cursor()
+            if self.logshow:
                 filename = self.logfile[self.select]
-                cur.execute('insert into log_select(name) values("%s")' % filename)
-        else:
-            data = get_all_log()[log_list[self.select]]
-            r = []
-            for i in data:
-                t = []
-                t.append(i)
-                r.append(t)
-
-            if self.select in self.mark_main:
-                self.mark_main.remove(self.select)
-                cur.executemany('DELETE FROM log_select WHERE name=?', r)
+                if self.select in self.mark[self.log_name]:
+                    self.mark[self.log_name].remove(self.select)
+                    cur.execute('delete from log_select where name="%s"' % filename)
+                else:
+                    self.mark[self.log_name].append(self.select)
+                    filename = self.logfile[self.select]
+                    cur.execute('insert into log_select(name) values("%s")' % filename)
             else:
-                self.mark_main.append(self.select)
-                cur.executemany('INSERT INTO log_select(name) VALUES(?)', r)
-        db.commit()
-        cur.close()
-        db.close()
+                data = get_all_log()[log_list[self.select]]
+                r = []
+                for i in data:
+                    t = []
+                    t.append(i)
+                    r.append(t)
+                if r and r[0][0]!='无':
+                    if self.select in self.mark_main:
+                        self.mark_main.remove(self.select)
+                        cur.executemany('DELETE FROM log_select WHERE name=?', r)
+                    else:
+                        self.mark_main.append(self.select)
+                        cur.executemany('INSERT INTO log_select(name) VALUES(?)', r)
+            db.commit()
+            cur.close()
+            db.close()
+            self.working = False
 
 
 if __name__ == '__main__':
